refactor: move status bar buttons for profile
